Fix console plugin circular dep causing bad CDN build
